move getCalcDisplay and SillyContext
