+++ libs/gi/page-archive/src/TabCharacter.tsx
@@ -12,8 +12,11 @@ import { useDBMeta } from '@genshin-optimizer/gi/db-ui'
 import { i18n } from '@genshin-optimizer/gi/i18n'
 import { getCharEle, getCharStat } from '@genshin-optimizer/gi/stats'
 import { ElementIcon } from '@genshin-optimizer/gi/svgicons'
-import { CharacterName, iconAsset } from '@genshin-optimizer/gi/ui'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
+import {
+  CharacterName,
+  SillyContext,
+  iconAsset,
+} from '@genshin-optimizer/gi/ui'
 import SearchIcon from '@mui/icons-material/Search'
 import StarRoundedIcon from '@mui/icons-material/StarRounded'
 import type { Palette } from '@mui/material'
