+++ libs/gi/page-settings/src/SillyCard.tsx
@@ -1,6 +1,6 @@
 import { DiscordIcon } from '@genshin-optimizer/common/svgicons'
 import { CardThemed } from '@genshin-optimizer/common/ui'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
+import { SillyContext } from '@genshin-optimizer/gi/ui'
 import { Android as AndroidIcon, Apple as AppleIcon } from '@mui/icons-material'
 import CheckBoxIcon from '@mui/icons-material/CheckBox'
 import CheckBoxOutlineBlankIcon from '@mui/icons-material/CheckBoxOutlineBlank'
