+++ libs/gi/page-team/src/CharacterDisplay/FormulaModal.tsx
@@ -12,11 +12,12 @@ import {
   AmpReactionModeText,
   DataContext,
   FormulaDataContext,
+  getCalcDisplay,
   getDisplayHeader,
   getDisplaySections,
 } from '@genshin-optimizer/gi/ui'
 import type { CalcResult } from '@genshin-optimizer/gi/uidata'
-import { getCalcDisplay, resolveInfo } from '@genshin-optimizer/gi/uidata'
+import { resolveInfo } from '@genshin-optimizer/gi/uidata'
 import type { DisplaySub } from '@genshin-optimizer/gi/wr'
 import CloseIcon from '@mui/icons-material/Close'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
