+++ libs/gi/page-team/src/CharacterDisplay/StatModal.tsx
@@ -16,8 +16,9 @@ import {
   FieldDisplayList,
   NodeFieldDisplay,
   StatEditorList,
+  getCalcDisplay,
 } from '@genshin-optimizer/gi/ui'
-import { getCalcDisplay, resolveInfo } from '@genshin-optimizer/gi/uidata'
+import { resolveInfo } from '@genshin-optimizer/gi/uidata'
 import type { ReadNode } from '@genshin-optimizer/gi/wr'
 import { allInputPremodKeys, uiInput as input } from '@genshin-optimizer/gi/wr'
 import BarChartIcon from '@mui/icons-material/BarChart'
