+++ libs/gi/ui/src/components/artifact/LocationFilterMultiAutocomplete.tsx
@@ -8,10 +8,10 @@ import {
 } from '@genshin-optimizer/gi/consts'
 import { useDBMeta, useDatabase } from '@genshin-optimizer/gi/db-ui'
 import { getCharEle } from '@genshin-optimizer/gi/stats'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
 import { Chip, Skeleton } from '@mui/material'
 import { Suspense, useCallback, useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
+import { SillyContext } from '../../context'
 import { CharIconSide } from '../character'
 
 export function LocationFilterMultiAutocomplete({
