+++ libs/gi/ui/src/components/character/CharacterCardPico.tsx
@@ -13,13 +13,13 @@ import {
   useDBMeta,
 } from '@genshin-optimizer/gi/db-ui'
 import { getCharStat } from '@genshin-optimizer/gi/stats'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
 import { ascensionMaxLevel } from '@genshin-optimizer/gi/util'
 import FavoriteIcon from '@mui/icons-material/Favorite'
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder'
 import { Box, CardActionArea, Typography } from '@mui/material'
 import type { MouseEvent, ReactNode } from 'react'
 import { useCallback, useContext } from 'react'
+import { SillyContext } from '../../context'
 import { iconAsset } from '../../util/iconAsset'
 
 export function CharacterCardPico({
