+++ libs/gi/ui/src/components/character/LocationAutocomplete.tsx
@@ -13,13 +13,13 @@ import {
   charKeyToLocGenderedCharKey,
 } from '@genshin-optimizer/gi/consts'
 import { useDBMeta, useDatabase } from '@genshin-optimizer/gi/db-ui'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
 import type { Variant } from '@genshin-optimizer/gi/wr'
 import BusinessCenterIcon from '@mui/icons-material/BusinessCenter'
 import type { AutocompleteProps } from '@mui/material'
 import { Box, Skeleton } from '@mui/material'
 import { Suspense, useCallback, useContext, useEffect, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
+import { SillyContext } from '../../context'
 import { CharIconSide } from './CharIconSideElement'
 type LocationAutocompleteProps = {
   location: LocationKey
