+++ libs/gi/ui/src/components/character/Trans.tsx
@@ -5,9 +5,9 @@ import {
   type LocationGenderedCharacterKey,
 } from '@genshin-optimizer/gi/consts'
 import { Translate } from '@genshin-optimizer/gi/i18n'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
 import { useContext } from 'react'
 import { useTranslation } from 'react-i18next'
+import { SillyContext } from '../../context'
 
 export function CharacterName({
   characterKey,
