+++ libs/gi/ui/src/components/character/card/CharacterCardHeader.tsx
@@ -8,12 +8,11 @@ import { characterAsset } from '@genshin-optimizer/gi/assets'
 import type { CharacterKey, ElementKey } from '@genshin-optimizer/gi/consts'
 import { useDBMeta } from '@genshin-optimizer/gi/db-ui'
 import { getCharEle, getCharStat } from '@genshin-optimizer/gi/stats'
-import { SillyContext } from '@genshin-optimizer/gi/uidata'
 import { ascensionMaxLevel } from '@genshin-optimizer/gi/util'
 import { input } from '@genshin-optimizer/gi/wr'
 import { Box, CardActionArea, Chip, Typography } from '@mui/material'
 import { useCallback, useContext } from 'react'
-import { DataContext } from '../../../context'
+import { DataContext, SillyContext } from '../../../context'
 import { iconAsset } from '../../../util/iconAsset'
 import type { RollColorKey } from '../../artifact'
 import { CharacterName } from '../Trans'
