+++ libs/gi/ui/src/context/SillyContext.tsx
@@ -1,7 +1,5 @@
 import { createContext } from 'react'
 
-// NOTE: SillyContext is only put in here due to its use in UIData.SourceDisplay.
-
 export type SillyContextObj = {
   silly: boolean
   setSilly: (s: boolean) => void
