+++ libs/gi/ui/src/context/index.ts
@@ -2,4 +2,5 @@ export * from './DataContext'
 export * from './FormulaDataContext'
 export * from './GraphContext'
 export * from './OptimizationTargetContext'
+export * from './SillyContext'
 export * from './SnowContext'
