+++ libs/gi/ui/src/hooks/useSilly.tsx
@@ -1,5 +1,5 @@
-import type { SillyContextObj } from '@genshin-optimizer/gi/uidata'
 import { useEffect, useMemo, useState } from 'react'
+import type { SillyContextObj } from '../context'
 const lsKey = 'silly_optimizer_enabled'
 
 export function useSilly(): SillyContextObj {
