+++ libs/gi/uidata/src/index.ts
@@ -1,2 +1,1 @@
-export * from './SillyContext'
 export * from './uiData'
