+++ libs/gi/uidata/src/uiData.tsx
@@ -1,24 +1,16 @@
-import { ColorText } from '@genshin-optimizer/common/ui'
 import {
   assertUnreachable,
   crawlObject,
   getUnitStr,
   layeredAssignment,
   objPathValue,
-  valueString,
 } from '@genshin-optimizer/common/util'
 import type {
   ArtifactSetKey,
   CharacterKey,
-  CharacterSheetKey,
   GenderKey,
-  WeaponKey,
 } from '@genshin-optimizer/gi/consts'
-import {
-  allArtifactSetKeys,
-  allCharacterSheetKeys,
-  allWeaponKeys,
-} from '@genshin-optimizer/gi/consts'
+import { allArtifactSetKeys } from '@genshin-optimizer/gi/consts'
 import { Translate } from '@genshin-optimizer/gi/i18n'
 import { KeyMap } from '@genshin-optimizer/gi/keymap'
 import { StatIcon } from '@genshin-optimizer/gi/svgicons'
@@ -29,7 +21,6 @@ import type {
   DisplaySub,
   Info,
   InfoExtra,
-  KeyMapPrefix,
   LookupNode,
   MatchNode,
   NumNode,
@@ -53,14 +44,12 @@ import {
   uiInput,
 } from '@genshin-optimizer/gi/wr'
 import type { ReactNode } from 'react'
-import { useContext } from 'react'
-import { SillyContext } from './SillyContext'
 
 /**
  * @deprecated
  * use CalcResult
  */
-export interface NodeDisplay<V = number> {
+interface NodeDisplay<V = number> {
   /** Leave this here to make sure one can use `crawlObject` on hierarchy of `NodeDisplay` */
   operation: true
   info: Info
@@ -81,16 +70,6 @@ export type CalcResult<V = number> = {
   }
   info: Info
   isEmpty: boolean
-}
-
-type CalcDisplay = {
-  name?: ReactNode
-  valueString: string
-
-  prec: number
-  formula: ReactNode
-  assignment?: ReactNode
-  formulas: ReactNode[]
 }
 
 export class UIData {
@@ -451,175 +430,6 @@ function accumulateInfo<V>(operands: CalcResult<V>[]): Info {
   return result
 }
 
-const displayKey = Symbol()
-export function getCalcDisplay(
-  node: CalcResult<number | string | undefined>
-): CalcDisplay {
-  if ((node as any)[displayKey]) {
-    return (node as any)[displayKey]
-  }
-  const result: CalcDisplay = _getCalcDisplay(node)
-  ;(node as any)[displayKey] = result
-  return result
-}
-function _getCalcDisplay(
-  node: CalcResult<number | string | undefined>
-): CalcDisplay {
-  const { info, value } = node
-  if (typeof value !== 'number') {
-    const str = value ?? ''
-    return {
-      prec: Infinity,
-      valueString: str,
-      formula: str,
-      formulas: [],
-    }
-  }
-
-  const infoExtra = resolveInfo(info)
-  const { name, prefix, source, variant } = infoExtra
-  const result = computeFormulaDisplay(node as CalcResult<number>, infoExtra)
-  if (node.info.path) {
-    const prefixDisplay = prefix && !source ? <>{subKeyMap[prefix]} </> : null
-    const sourceDisplay = <SourceDisplay source={source} />
-    const nameVariant = variant && variant === 'invalid' ? undefined : variant
-    result.name = (
-      <>
-        <ColorText color={nameVariant}>
-          {prefixDisplay}
-          {prefixDisplay && ' '}
-          {name}
-        </ColorText>{' '}
-        {sourceDisplay}
-      </>
-    )
-
-    if (node.meta.op !== 'const')
-      result.assignment = (
-        <div className="formula">
-          {result.name} <ColorText color="info">{result.valueString}</ColorText>{' '}
-          = {result.formula}
-        </div>
-      )
-
-    if (node.info.pivot || node.meta.op === 'const') {
-      result.prec = Infinity
-      result.formula = (
-        <>
-          <span style={{ fontSize: '85%' }}>{result.name}</span>{' '}
-          {result.valueString}
-        </>
-      )
-
-      if (result.assignment) {
-        result.formulas = [result.assignment, ...result.formulas]
-        delete result.assignment
-      }
-    }
-  }
-  return result
-}
-
-function computeFormulaDisplay(
-  node: CalcResult<number>,
-  info: Info & InfoExtra
-): CalcDisplay {
-  const details = {
-    add: { head: '', sep: ' + ', tail: '', prec: 1 },
-    mul: { head: '', sep: ' * ', tail: '', prec: 2 },
-    max: { head: 'Max(', sep: ', ', tail: ')', prec: Infinity },
-    min: { head: 'Min(', sep: ', ', tail: ')', prec: Infinity },
-  } as const
-
-  const { op, ops } = node.meta
-  const result: CalcDisplay = {
-    prec: Infinity,
-    formula: '',
-    valueString: valueString(node.value, info.unit, info.fixed),
-    formulas: [...new Set(ops.flatMap((op) => getCalcDisplay(op).formulas))],
-  }
-  const components: ReactNode[] = []
-
-  function addComponents(node: CalcResult<number>, p: number) {
-    const display = getCalcDisplay(node)
-    if (p > display.prec) components.push('(')
-    components.push(display.formula)
-    if (p > display.prec) components.push(')')
-  }
-
-  switch (op) {
-    case 'const':
-      components.push(result.valueString)
-      break
-    case 'add':
-    case 'mul':
-    case 'min':
-    case 'max': {
-      if (ops.length === 1) result.prec = getCalcDisplay(ops[0]).prec
-      else result.prec = details[op].prec
-      const { head, sep, tail, prec } = details[op]
-      components.push(head)
-      ops.forEach((op, i) => {
-        if (i !== 0) components.push(sep)
-        addComponents(op, prec)
-      })
-      components.push(tail)
-      break
-    }
-    case 'sum_frac': {
-      result.prec = details.mul.prec
-      addComponents(ops[0], details.mul.prec)
-      components.push(' / (')
-      addComponents(ops[0], details.add.prec)
-      components.push(' + ')
-      addComponents(ops[1], details.add.prec)
-      components.push(')')
-      break
-    }
-    case 'res': {
-      const [preRes] = ops
-      if (preRes.value >= 0.75) {
-        result.prec = details.mul.prec
-        components.push('1 / (1 + 4 * ')
-        addComponents(preRes, details.mul.prec)
-        components.push(')')
-      } else if (preRes.value >= 0) {
-        result.prec = details.add.prec
-        components.push('1 - ')
-        addComponents(preRes, details.add.prec)
-      } else {
-        result.prec = details.add.prec
-        components.push('1 - ')
-        addComponents(preRes, details.mul.prec)
-        components.push(' / 2')
-      }
-    }
-  }
-
-  components.filter((c) => c)
-  result.formula = (
-    <>
-      {components.map((x, i) => (
-        <span key={i}>{x}</span>
-      ))}
-    </>
-  )
-
-  return result
-}
-
-const subKeyMap: Record<KeyMapPrefix, string> = {
-  default: 'Default',
-  base: 'Base',
-  total: 'Total',
-  uncapped: 'Uncapped',
-  custom: 'Custom',
-  char: 'Char.',
-  art: 'Art.',
-  weapon: 'Weapon',
-  teamBuff: 'Team',
-}
-
 export function resolveInfo(info: Info): Info & InfoExtra {
   // Make sure not to override any `Info` (those should be calculated
   // in `UIData`) so we only write to `extra` and combine later
@@ -643,37 +453,6 @@ export function resolveInfo(info: Info): Info & InfoExtra {
   }
 
   return Object.keys(extra).length ? { ...info, ...extra } : info
-}
-function SourceDisplay({ source }: { source: string | undefined }) {
-  const { silly } = useContext(SillyContext)
-  if (!source) return null
-  if (allArtifactSetKeys.includes(source as ArtifactSetKey))
-    return (
-      <ColorText color="secondary">
-        {' '}
-        (<Translate ns="artifactNames_gen" key18={source} />)
-      </ColorText>
-    )
-  if (allWeaponKeys.includes(source as WeaponKey))
-    return (
-      <ColorText color="secondary">
-        {' '}
-        (<Translate ns="weaponNames_gen" key18={source} />)
-      </ColorText>
-    )
-  if (allCharacterSheetKeys.includes(source as CharacterSheetKey))
-    return (
-      <ColorText color="secondary">
-        {' '}
-        (
-        <Translate
-          ns={silly ? 'sillyWisher_charNames' : 'charNames_gen'}
-          key18={source}
-        />
-        )
-      </ColorText>
-    )
-  return null
 }
 
 function mergeInfo(base: Info, override: Info): Info {
